+++ chat-application/lib/l10n/flyer_l10n.dart
