+++ chat-application/lib/widgets/chat/widgets/error.dart
