+++ chat-application/lib/widgets/chat/widgets/switch_group.dart
